enhancements

+++ Katiba55.API/Dtos/Items/ItemDetailedDto.cs
@@ -8,11 +8,11 @@
         public DateTime? EstimatedEndDate { get; set; }
         public DateTime? ActualStartDate { get; set; }
         public DateTime? ActualEndDate { get; set; }
-        public decimal? TotalValue { get; set; }
-        public decimal? ExecutedValue { get; set; }
-        public decimal? RemainingValue { get; set; }
-        public decimal? RelativeWeightPercent { get; set; }
-        public decimal? ExecutionPercent { get; set; }
+        public double? TotalValue { get; set; }
+        public double? ExecutedValue { get; set; }
+        public double? RemainingValue { get; set; }
+        public double? RelativeWeightPercent { get; set; }
+        public double? ExecutionPercent { get; set; }
         public DateTime? ExecutionDate { get; set; }
         public string ExecutionStatus { get; set; }
         public string? Notes { get; set; }
