+++ Katiba55.API/Dtos/Projects/ProjectDetailedDto.cs
@@ -8,8 +8,8 @@ namespace Katiba55.API.Dtos.Projects
         public string Name { get; set; }
         public string? ExecutingSide { get; set; }
         public string? BenefitingSide { get; set; }
-        public decimal? EstimatedCost { get; set; }
-        public decimal? FinancialAllocation { get; set; }
+        public double? EstimatedCost { get; set; }
+        public double? FinancialAllocation { get; set; }
         public DateTime? EstimatedStartDate { get; set; }
         public DateTime? EstimatedEndDate { get; set; }
         public DateTime? ActualStartDate { get; set; }
@@ -17,7 +17,7 @@ namespace Katiba55.API.Dtos.Projects
         public string? Address { get; set; }
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
-        public decimal? ExecutionPercent { get; set; }
+        public double? ExecutionPercent { get; set; }
         public DateTime? ExecutionDate { get; set; }
         public string ExecutionStatus { get; set; }
