+++ Katiba55.API/Dtos/Projects/UpdateProjectDto.cs
@@ -7,8 +7,8 @@ namespace Katiba55.API.Dtos.Projects
         public string Name { get; set; }
         public string? ExecutingSide { get; set; }
         public string? BenefitingSide { get; set; }
-        public decimal? EstimatedCost { get; set; }
-        public decimal? FinancialAllocation { get; set; }
+        public double? EstimatedCost { get; set; }
+        public double? FinancialAllocation { get; set; }
         public DateTime? EstimatedStartDate { get; set; }
         public DateTime? EstimatedEndDate { get; set; }
         public DateTime? ActualStartDate { get; set; }
@@ -17,7 +17,7 @@ namespace Katiba55.API.Dtos.Projects
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
         public int SupervisorId { get; set; }
-        public decimal? ExecutionPercent { get; set; }
+        public double? ExecutionPercent { get; set; }
         public DateTime? ExecutionDate { get; set; }
         [EnumDataType(typeof(ExecutionStatus))]
         public string ExecutionStatus { get; set; }
