+++ Katiba55.API/Dtos/Works/WorkWithItemsBriefDto.cs
@@ -6,7 +6,7 @@ namespace Katiba55.API.Dtos.Works
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal? ExecutionPercent { get; set; }
+        public double? ExecutionPercent { get; set; }
         public DateTime? ExecutionDate { get; set; }
         public string ExecutionStatus { get; set; }
