+++ Katiba55.API/Entities/Project.cs
@@ -5,8 +5,8 @@
         public string Name { get; set; }
         public string? ExecutingSide { get; set; } 
         public string? BenefitingSide { get; set; } 
-        public decimal? EstimatedCost { get; set; } 
-        public decimal? FinancialAllocation { get; set; }
+        public double? EstimatedCost { get; set; } 
+        public double? FinancialAllocation { get; set; }
         public DateTime? EstimatedStartDate { get; set; }
         public DateTime? EstimatedEndDate { get; set; }
         public DateTime? ActualStartDate { get; set; }
@@ -15,7 +15,7 @@
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
 
-        public decimal? ExecutionPercent { get; set; }
+        public double? ExecutionPercent { get; set; }
         public DateTime? ExecutionDate { get; set; }
         public ExecutionStatus ExecutionStatus { get; set; }
