+++ Katiba55.API/Entities/WorkExecutionHistory.cs
@@ -2,7 +2,7 @@
 {
     public class WorkExecutionHistory: BaseEntity
     {
-        public decimal Percentage { get; set; }
+        public double Percentage { get; set; }
         public DateTime Date { get; set; }
 
         public int WorkId { get; set; }
